requirements.txt
Django==1.9
django-appconf==1.0.1
django-autoslug==1.9.3
django-autoslug-field==0.2.3
django-bootstrap3==7.0.0
django-cked==0.1.2
django-ckeditor==5.0.3
django-compressor==2.0
django-debug-toolbar==1.4
django-extensions==1.6.1
django-flatpages-tinymce==0.1.1
django-registration==2.0.4
django-tinymce==2.2.0
docutils==0.12
easy-thumbnails==2.3
MySQL-python==1.2.5
Pillow==3.1.1
rcssmin==1.0.6
rjsmin==1.0.12
six==1.10.0
South==1.0.2
sqlparse==0.1.18
suit==2.0.2
wheel==0.24.0
